[any message]

+++ src/main/java/MyScanner.java
@@ -140,7 +140,7 @@ public class MyScanner {
 
             //If character not in enum
             default:
-                System.err.println("[line " + curLine + "] Error: Unexpected character: " + (int)c);
+                System.err.println("[line " + curLine + "] Error: Unexpected character: " + c);
                 errCode = 65;
 
         }
